src/requirements.txt
bcrypt==4.0.1
boto3==1.26.9
botocore==1.29.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
pycparser==2.21
Pillow==9.3.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
s3transfer==0.6.0
urllib3==1.26.12
Werkzeug==2.2.2
